Specifiation changes: RecaptchaSettings.cs have been moved to it's own settings folder to better organize the application flow. RecaptchaShape now uses IOptions<Recaptcha> to allow post configuration of Recaptcha's builder extension method to allow envrionment variables use case. OrchardCoreBuilderExtensions.cs has been added as part of the extension method for RecaptchaSettings.
In  the orchardcore drivers, the adjusted namespace to account for RecaptchaSettings.cs new position have been set for the drivers, forms, and the startup.cs file itself.
