+++ src/OrchardCore.Cms.Web/Program.cs
@@ -6,6 +6,7 @@ builder.Host.UseNLogHost();
 
 builder.Services
     .AddOrchardCms()
+    .ConfigureReCaptchaSettings()
     .AddSetupFeatures("OrchardCore.AutoSetup");
 
 var app = builder.Build();
