+++ src/OrchardCore.Modules/OrchardCore.ReCaptcha/Drivers/ReCaptchaLoginFormDisplayDriver.cs
@@ -1,9 +1,9 @@
 using System.Globalization;
 using OrchardCore.DisplayManagement.Handlers;
 using OrchardCore.DisplayManagement.Views;
-using OrchardCore.ReCaptcha.Configuration;
 using OrchardCore.Settings;
 using OrchardCore.Users.Models;
+using OrchardCore.ReCaptcha.Settings;
 
 namespace OrchardCore.ReCaptcha.Drivers;
 
@@ -19,11 +19,6 @@ public sealed class ReCaptchaLoginFormDisplayDriver : DisplayDriver<LoginForm>
     public override async Task<IDisplayResult> EditAsync(LoginForm model, BuildEditorContext context)
     {
         var settings = await _siteService.GetSettingsAsync<ReCaptchaSettings>();
-
-        if (!settings.ConfigurationExists())
-        {
-            return null;
-        }
 
         return Dynamic("ReCaptcha", (m) =>
         {
